==> globalprep/spp/v2021/_setup/2a_set_up_regional_codes.Rmd <==
---
title: 'Set up IUCN marine species ranges by regional assessment'
author: "*Compiled on `r date()` by `r Sys.info()['user']`*"
output: 
  html_document:
    code_folding: hide
    toc: true
    toc_depth: 3
    toc_float: yes
    number_sections: true
    theme: cerulean
    highlight: haddock
    includes: 
      in_header: '../../../../workflow/templates/ohi_hdr.html'
  pdf_document:
    toc: true
---

``` {r setup, echo = TRUE, message = FALSE, warning = FALSE}

knitr::opts_chunk$set(fig.width = 6, fig.height = 4, fig.path = 'Figs/',
                      echo = TRUE, message = FALSE, warning = FALSE)


library(dplyr)
library(rgdal)
library(raster)
library(here)
#library(ggridges)
library(ggplot2)

source(here('workflow/R/common.R'))

### provenance tracking
#library(provRmd); prov_setup()

goal     <- 'spp'
scenario <- 'v2021'
dir_goal <- here('globalprep', goal, scenario)

### goal specific folders and info
dir_anx  <- file.path(dir_M, 'git-annex/globalprep')
dir_setup   <- file.path(dir_goal, '_setup')
dir_goal_anx <- file.path(dir_anx, goal, scenario, 'spp_risk_dists')

source(file.path(dir_setup, 'common_fxns.R'))

```

# Summary

Examine IUCN Red List regional assessments for extinction risk categories.

# Data Sources

### IUCN Red List

# Methods: Geographic scope of regional assessments

For each regional assessment, spatially classify by [Large Marine Ecosystems](https://upload.wikimedia.org/wikipedia/commons/4/4d/Global_map_of_large_marine_ecosystems.jpg) (other potential options: EEZs and/or FAO Major Fishing Areas).  Use the [https://en.wikipedia.org/wiki/United_Nations_geoscheme](UN geoscheme) to roughly identify subregional zones.  Note there may be some overlaps.

## Europe ('europe'):

from European marine fishes: 

> The geographic scope of the European Red List assessment
encompasses the Mediterranean Sea, the Black Sea, the
Baltic Sea, the North Sea and/or the European part of
the Atlantic Ocean (i.e., the territorial waters and the
Exclusive Economic Zones (EEZs) of all European
countries in the eastern part of the Atlantic Ocean,
also including the EEZs of the Macaronesian islands
belonging to Portugal and Spain), the North Sea and
the Northeastern Atlantic Ocean, excluding the EEZs of
Greenland, Morocco and Western Sahara.

Options: 

* LME 20-26 (26 = Mediterranean) and 59 (Iceland)?
* MEOW 2 (Northern Iceland/Svalbard), [21:27, 30:31, 34:36] (Mediterranean), 44 (Black Sea)
* Alternately: Use EEZs for all countries listed as Europe by UN classification

## Northern Africa 

* LME 33 (Red Sea), 26 (Mediterranean), 27 (West Coast of North Africa)
* MEOW 87-88 (Red Sea), [32, 33, 35, 36] (Mediterranean), 28, 29 (Lusitanean upwelling)

## Southern Africa 

* LME 29 (Western Southern Africa) and 30 (Madagascar)
* MEOW 192:194 (Temperate southern Africa)

## Eastern Africa 

* LME 30 (Madagascar) and 31 (Horn of Africa)
* MEOW 94:102 (Western Indian Ocean)

## Western Africa   

* LME 28
* MEOW 79:86 (West African Transition, Gulf of Guinea)

## Central Africa  

* LME 28, 29
* If this is truly "Central" then no marine anyway! ignore

## Northeastern Africa 

* LME 32, 33
* MEOW 87:93 (Red Sea, Gulf of Aden, Somali/Arabian)

## Pan-Africa 

* LME 26-33
* MEOW 28, 29, 32, 33, 35, 36 (Med and N), 
    79:86 (West), 87:89, 93 (NE), 94:102 (E), 192:194 (S)

## Gulf of Mexico 

* LME 5
* MEOW 43 (N Gulf), 68:70 (S Gulf)

## Mediterranean

* LME 26 and 62 for Mediterranean and Black Seas
* MEOW 30:36, 44

## Persian Gulf

* LME 32 (NOTE: this encompasses the entire Arabian Sea)
* MEOW 90:91

``` {r code MEOWs}

rgn_to_meow <- c('europe'             = paste(c(2,21:27, 30:31, 34:36, 44), collapse = ','),
                'northern_africa'     = paste0(c(87:88, 32, 33, 35, 36, 28, 29), collapse = ','),
                'western_africa'      = paste0(79:86, collapse = ','), 
                # 'central_africa'      = '28,29',
                'northeastern_africa' = paste0(87:93, collapse = ','),
                'southern_africa'     = paste0(192:194, collapse = ','),
                'eastern_africa'      = paste0(94:102, collapse = ','),
                'pan-africa'          = paste0(c(28, 29, 32, 33, 35, 36, 79:89, 93:102, 192:194), 
                                                 collapse = ','),
                'gulf_of_mexico'      = paste0(c(43, 68:70), collapse = ','),
                'mediterranean'       = paste0(c(30:36, 44), collapse = ','), 
                'persian_gulf'        = paste0(90:91, collapse = ','),
                'global'              = paste0(1:232, collapse = ','))

rgn_to_meow_df <- data.frame(iucn_rgn = names(rgn_to_meow),
                             meow_id  = rgn_to_meow) %>%
  mutate(meow_id = str_split(meow_id, ',')) %>%
  unnest(meow_id) %>%
  group_by(iucn_rgn) %>%
  mutate(meow_id = as.integer(meow_id),
         priority = n()) %>%
  arrange(priority)

write_csv(rgn_to_meow_df, file.path(dir_goal, sprintf('_spatial/iucn_rgn_to_meow.csv')))

```

``` {r code LMEs}

rgn_to_lme <- c('europe'              = '20,21,22,23,24,25,26,59',
                'northern_africa'     = '26,27,33',
                'western_africa'      = '28', 
                'central_africa'      = '28,29',
                'northeastern_africa' = '32,33',
                'southern_africa'     = '29,30',
                'eastern_africa'      = '30,31',
                'pan-africa'          = '26,27,28,29,30,31,32,33',
                'gulf_of_mexico'      = '5',
                'mediterranean'       = '26,62', 
                'persian_gulf'        = '32',
                'global'              = paste0(1:66, collapse = ','))

rgn_to_lme_df <- data.frame(iucn_rgn = names(rgn_to_lme),
                            lme_id = rgn_to_lme) %>%
  mutate(lme_id = str_split(lme_id, ',')) %>%
  unnest(lme_id) %>%
  group_by(iucn_rgn) %>%
  mutate(lme_id = as.integer(lme_id),
         priority = n()) %>%
  arrange(priority)

write_csv(rgn_to_lme_df, file.path(dir_goal, sprintf('_spatial/iucn_rgn_to_lme.csv')))

```

Where an LME is assigned to multiple regions, use a prioritization then averaging method:

* Prioritize by the region comprising the smallest number of LMEs; e.g. Mediterranean has 1 LME, North Africa has 3, Europe has 8; they all overlap in LME 26.  So for the score for LME 26 for a species assessed in all three regions, keep the score for the Mediterranean.
* If after prioritizing, multiple regions still overlap in an LME, take the average score of the overlapping regional assessments for that LME.

-----

``` {r prov_footer, results = 'asis'}

#prov_wrapup(commit_outputs = FALSE)

```
